+++ src/components/features/fasting/index.ts
@@ -1,0 +1,8 @@
+/**
+ * Fasting feature components exports
+ */
+
+export { FastingHeader } from "./FastingHeader";
+export { FastingTimeInput } from "./FastingTimeInput";
+export { FastingTimeline } from "./FastingTimeline";
+export { FastingPhaseCard } from "./FastingPhaseCard";