+++ src/types/index.ts
@@ -4,6 +4,7 @@
 
 export * from "./auth.types";
 export * from "./sleep.types";
+export * from "./fasting.types";
 export * from "./caffeine.types";
 export * from "./todo.types";
 export * from "./notification.types";